Working Json Tile Writing
