+++ Assets/_Laz/Scripts/ScriptableObject/Laz/MovementProperties.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1d2ac3683cedc074d834723c00447237
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
