+++ Assets/_Laz/Scripts/Tools/Writer.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 503ebcfae40fd284f905570a3b854b28
+guid: ac35f78505d512d408dc10cb1a3a2bef
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
